UPBGE: Remove KX_GameObject::UpdateMaterialData.
Previously the material action/IPO was updating materials from the
KX_GameObject with the material name as identifier.

To simplify the code the UpdateMaterialData function was removed
and the material name was replaced by a pointer to RAS_IPolyMaterial
in KX_MaterialIpoController. But it assume that we need to find
the material from BL_Action, to doing it we use a scene converter
convertion function : blende rmaterial to rastrizer material.