UPBGE: Implement physics mesh shape duplication.
As meshes can be duplicated with LibNew, meshes physics shape should too.
To ask a duplication the user set the argument "dupli" to True in
KX_GameObject.reinstancePhysicsMesh. And then internaly it check if the
shape as at least two users, and if yes we make a copy of the
CcdShapeConstructionInfo and update it with the passed mesh.

Questions:
- Why not use a function only to duplicate the physics mesh shape ?
  - It's not possible because the duplicated physics shape should be
    update just after its duplication to be valid. We can keep the a
    pointer to the old bullet shape but it cause issues. The worse
    case is if the shape is replicated (it keep pointer to old bullet shape)
    and then the original shape is deleted (the game object is freed).
    Here we're sure to cause a crash.