UPBGE: Fix component initialization without children.
Previously when a parent object was duplicated the components in it was initialized before
add the children in this parent object. Then it was impossible to do something with children.

To solve this we call InitComponents at the end of AddReplicaObject and also call InitComponents
in all recursive children, the user should not respect an initialization order (e.g parent
and after children) but we ensure that every children are here..

The InitComponent call in ProcessReplica is tehn removed.