UPBGE: Implement KX_2DFilter python proxy based on BL_Shader.
Previously the 2D filters was alterated only through a 2D filter actuator.
This limited the possibility of the user to change the shader in game or
to update a particular uniform type (e.g matrix or vector or defined uniform)
or from outside of the object that owned the actuator.

The new way is to use two new classes.
The first is KX_2DFilter which is a class based on RAS_2DFilter for the
filter usage and BL_Shader for the python API to alterate the shader.
The second class is KX_2DFilterManager that only exist to override the
function NewFilter to instance KX_2DFilter instead of RAS_2DFilter and
also to expose trivial filter function to the user to add, remove and
find a 2D filter.

This last class is instancied in KX_Scene like the previous RAS_2DFilterManager
and accesible in python with scene.filterManager. And offers to the user
the functions:
scene.filterManager.getFilter(index)
scene.filterManager.addFilter(index, type, fragmentProgram)
scene.filterManager.removeFilter(index)

About the KX_2DFilter python API a new function is introduced:
setTexture(index, bindCode) to bind a texture "bindCode" in slot "index"
before draw the filter, This function requierd to use setSampler too.

The RAS_2DFilter was a bit refactored to allow changing of shader after its
contruction by overriding the LinkProgram function to set a boolean variale
m_uniformInitialized to false, then this variable will allow later to
parse the shader and recreate the textures.
Also the draw functions was shared with the rasterizer function previously
used to draw the world background renamed to RAS_IRasterizer::DrawOverlayPlane().
Which request to separate the depth func set out from it in
KX_WorldInfo::RenderBackground.

The customs textures was allowed for 2D filter and alterated thanks to
setTexture python function, but as this function allow textures only from 0
to 7 then the built-in textures (bgl_ prefixed) must be bound in slot 8 to 10.

List of commits:
- Implementation of KX_Filter proxy to set uniforms
- clean up and add GPL license in KX_2DFiltermanager.h/.cpp
- Rebase KX_2DFilter upon BL_Shader.
- Add dummy KX_2DFilterManager python functions : getFilter/addFilter/removeFilter.
- Add scene.filterManager in python
- UPBGE: Add pytohn attribut [BL_Shader/KX_2DFilter].enabled = bool.
- Remove unused pass index from RAS_2DFilter, pass index is managed only in the RAS_2DFilterManager
- Rename enabled BL_Shader attribute to "enable"
- Revert "Rename enabled BL_Shader attribute to "enable""
- Remove unused RAS_2DFilterManager::PrintShaderError.
- Use BLI_utildefines include for STRINGIFY macro.
- Parse uniforms, create inexisting texture and compute texture offset for every LinkProgram call.
- Implement filter textures + unbind filter textures.
- Make KX_2DFilter::LinkProgram explicitly using RAS_2DFilter::LinkProgram.
- Reset textures bindcode to 0 when recreating shader.
- Use RAS_IRasterizer OpenGL call wrapper functions.
- Implement common RAS_IRasterizer::DrawOverlayPlane function.
- Allow add filter from python with function addFilter(index, type, [fragment]).
- Remove unused variable in RAS_2DFilterData and check bind code in KX_2DFilter.setTexture.
- Add documentation.
- Minors fix about unfreed texture and licence blocks.