UPBGE: Refactor BL_Shader in RAS_Shader.
Previously the BL_Shader class contained all the functions useful to create
a custom shader and bind uniform/attribut to it and provide a python proxy.
This class was in Ketsji directory and created two issues:
- The code called OpenGL calls in Ketsji which should be disallowed.
- The code linked the python proxy to the helper function to creating shaders.

To solve this the old BL_Shader class was split in two classes with both a
different goal:

RAS_Shader:
This class is located directly in Rasterizer/ and contains all the helper
function for shaders and uniforms.

BL_Shader:
This class is now only the python proxy definition upon RAS_Shader.

Since all shader functions are moved in RAS_Shader, then RAS_2DFilter can
reuse they. To do it the RAS_2DFilter simply inherite of RAS_Shader which
obligated the RAS_Shader class to be not anymore mesh specific. Only the
Update(...) functions was mesh specific to get the mesh matrix, the
RAS_MeshSlot argument of this function is now replaced by a MT_Matrix4x4.

In the same way RAS_2DFilter is refactored, first to remove duplicated
attribute and functions with RAS_Shader and to use SetUniform functions, also
this functions is changed to support vector, matrix, float arrays.

Commit list:
- Copy BL_Shader to RAS_Shader
- Readd RAS_Shader::Update
- Inherit BL_Shader from RAS_Shader.
- Rebase RAS_2DFilter upon RAS_Shader.
- Remove m_initialized, m_error, m_enabled in RAS_2DFilter.
- Remove extra debug message for unfound uniform.
- Remove [Enable/Disable]FilterPass.
- Use enums for uniforms and textures types.
- Use RAS_Shader::SetUniform instead of glUniform1fARB
- Remove glew-mx.h from BL_Shader, add ValidateProgram and DeleteShader. (+ clean includes)
- Remove commented code in RAS_2DFilter
- Rename RAS_VertexFilter.h to RAS_VertexShader2DFilter.h
- Add and fix support for float and int vector arrays.
- Move SORT_UNIFORMS in RAS_Shader.h.
- Use STR_String instead of const char * to sotre vertex and fragment program.
- Remove mPass from RAS_Shader.