UPBGE: Allow writable texture.bindCode
Changing the bind code is useful to replace a texture by a shadow texture.
In this case : texture.bindCode = lamp.shadowBindId

A bug was found in the same time : setting a custom shader recreate material
textures because of the second call to InitTextures(). This call is removed
for the moment but can be related to libload issue:
Load an object with a component that alterate a material before merge the
used material in the proper scene (the blender shader and the blender
textures will be not here).