UPBGE: Implement vehicle constraint ray cast mask.
Previously it was impossible to allow the wheel to use a physics controller
because its shape was frequently hit by the ray.

To support it the btVehicleRaycaster is inherited to BlenderVehicleRaycaster
which use a m_mask attribute checked with a & operator to the hit object
collision group.

From the python API a new attribute is introduced : KX_VehicleWrapper.rayMask,
it reacts like KX_GameObject.collisionMask.