UPBGE: Joystick refactor
This Joystick refactor comes from the issues resulted from the migration to SDL2 library.

SDL2 uses Xinput library for Windows and it works different from old mm and dinput Windows libraries and additionally the Windows behaviour is
not consistent with linux/MacOS behaviour.

To mitigate this issue SDL2 implements a new Game Controller API. This Game Controller API changes totally the philosophy of Joystick support
on PC comparing it to Joystick support on Consoles (PS4, Xbox, etc). That is: It defines that the standard it is the XBox 360 controller and It
needs that all the Joysticks are mapped against it. For this purpose, this new API uses a intermediate file (SCA_JoystickMappingdb.h) where the
joysticks are mapped to follow a similar configuration than Xbox 360 game controller. This file is a slightly different version of file from
https://github.com/gabomdq/SDL_GameControllerDB/blob/master/gamecontrollerdb.txt (which it is the community effort to map every Joystick
against XBox360 one).

Additionally, If a Joystick is not mapped there are several tools availables (Steam Big Mode, controllermapping in SDL2 test sources and
antimicro) to perform this mapping.

Then, To clarify a little how it works now regarding our unintuitive Joystick sensor:
A Game Controller is composed of:

6 axis availables (in order from 0 to 5):

    AXIS_LEFTSTICK_X, AXIS_LEFTSTICK_Y, AXIS_RIGHTSTICK_X, AXIS_RIGHTSTICK_Y, AXIS_TRIGGERLEFT and AXIS_TRIGGERRIGHT.

15 buttons availables (in order from 0 to 14):

    BUTTON_A, BUTTON_B, BUTTON_X, BUTTON_Y, BUTTON_BACK, BUTTON_GUIDE, BUTTON_START, BUTTON_LEFTSTICK, BUTTON_RIGHTSTICK, BUTTON_LEFTSHOULDER,
BUTTON_RIGHTSHOULDER, BUTTON_DPAD_UP, BUTTON_DPAD_DOWN, BUTTON_DPAD_LEFT and BUTTON_DPAD_RIGHT.

The Joystick sensor UI has been reworked to follow these new axis and buttons set-up.

Also, to exploit new SDL2 hotpluggin capabilities the Joystick code has been refactored to create or destroy a Game Controller when it receives
a SDL_CONTROLLERDEVICEADDED or SDL_CONTROLLERDEVICEREMOVED event only, instead of to create all joysticks at start. This way the UPBGE support
Joystick hotpluggin now.

I have tested this patch in Linux/Windows with a Xbox 360 controller and a PS2 clone. Xbox 360 and PS2 clone worked out the box and the
behaviour was consistent between linux and windows (not as now).

WARNING:
This patch is NOT BACKWARD COMPATIBLE due to new API. Between other things the new API has not "hat" concept and it processes "hat/dpad"
control as any other button (BUTTON_DPAD_UP, BUTTON_DPAD_DOWN, BUTTON_DPAD_LEFT and BUTTON_DPAD_RIGHT). Also, it implement Shoulder trigger
buttons (left/right) as independent axis.

Moreover, to implement a "do_version" it is very hard because we would have to differentiate between MacOS&Linux and Windows, and between
Joysticks and Game Controllers. The only way to make it compatible with previous versions is to compile SDL2 disabling XINPUT and enabling
MM/DINPUT but this way we would loose hottpluggin, haptic (vibration) and standatization (at last) features.