in ge_filter, replace setTexture(m_textures index, bindcode)
with setTexture(textureName, textureBindCode)

To make textures available in 2D filters, Panzergame made a list of 8
textures slots available called m_textures[8]. To assign a texture into each
of these textures slots, we have to replace its default value (0) with the
bindcode of an existing texture:

m_textures[0] = 0 becomes m_textures[0] = bindcode of an existing texture

The previous method setTexture was setTexture(m_textures index, bindcode).
This way, m_textures[index] was equal to bindcode.

After we have assigned a bindcode to m_textures[index] we had to call:
filter.setSampler to make uniform sampler2D texturename available in the
shader like in BL_Shader.

I propose to merge the 2 steps: setTexture + setSampler:

setTexture("texturename", texturebindcode)

In the setTexture method I implemented, we check in a for loop if
m_textures[i] == 0 (default value). If so, this means the m_textures[i] is
available to assign a bindcode. So we assign a bindcode to it, and then we
set the uniform directly in the shader.