UPBGE: Fix bug when using glActiveTexture(bgl) in python script
Set the current active OpenGL texture to the first texture
and bind a null texture in this slot at the end of frame.
This function must be used very carfully, normally only after
that the user palyed with glActiveTexture and to make sure that
it will not break the render.
Only the first slot is affected all texture in greater slot are
not affected but just unused as default.