UPBGE: Fix 2DFilter performance issue and restore old background
in RenderBackground, the DrawOverlayPlane function use performances.
When an option blend sky, paper sky or real sky is checked, we render the
viewport sky. but if none of this option is checked, we render the old
background. The old background API is restored. (deprecation warnings for
backgroundColor are removed).
If the .blend has no world, we render glClearColor(0.224, 0.224, 0.224,
1.0) (default grey in the viewport).
when we don't use the DrawOverlayPlane function, the 2D Filter
performances are restored as before.

We could see if we can improve DrawOverlayPlane function (used both in
background rendering and 2DFilters rendering) to earn some fps. After
this, all the users that use unreal ungine, unity, cryengine and godot
will move to upbge. It's a fact