UPBGE: Fix SSS with geometry instancing.
Previously the conversion material to shader function used for the geometry
instancing was adding a linear to srgb call at the end of the shader contrary
to the reguler material conversion function.

To avoid any new similar issue due to the non-update of the changes in the
first function to its copy, we merge the both function and simply add a boolean
argument at the end "is_instancing" to create a geometry instacing material
and add it in the right gpu material list.