UPBGE: Fix render to texture background rendering when no world
When there is no world, the m_horizonColor values from KX_WorldInfo are not
available to render the background in ImageRender.

So before rendering the background in ImageRender we check if there is a
world. If not, we call setClearColor to set the defaults viewport
background colors and we don't use RenderBackground function which cause
a GLSL error.