UPBGE: Add a new method to setUniformTexture in KX_2DFilter
I think the system with setTexture then setSampler is too complex to
understand for the user. But as setTexture method could be used to assign
a bindcode to m_texture[index] without having to setSampler, I propose to
add another method in KX_2DFilter to set uniform as we did usually in
BL_Shader methods: setUniform...(name, value). I think this is more
understandable for the end user.
The setUniformTexture(textureName, textureBindCode, index) method sets
uniform sampler2D texturename with the bindcode of the texture we want to
bind in the 2D Filter. Index argument is optional: It allows the user to
choose at which emplacement of m_textures list he wants to bind his
texture. Like that, if all the emplacement in m_textures list are already
filled, the user can choose to overwrite an existing emplacement with his
texture bindcode.
If all emplacements in m_textures list are already filled, the user can't
bind additional texture in his 2D filter. A warning message inform him
that all the available m_textures slots are already filled with a
bindcode, so that he has to use optional index argument to overwrite
m_textures[index] with his bindcode.