UPBGE: Implement python loop script override in blenderplayer.
This feature is only for debug purpose. It allow the user to set
an override python main loop script when calling the blenderplayer.

This script is used by passing it's filename after the -p argument,
e.g:
blenderplayer -p pythonloop.py file.blend

The script used for main loop in general is get by calling
LA_Launcher::GetMainLoopPythonCode. This function overrided in
LA_PlayerLauncher to return the script passed in -p if it exists
instead of the scene property.