UPBGE: Remove SG_QList from PyObjectPlus.
Previously this class was used only for implicit list maagement for logic bricks
and armature channels/contraints.
This class already caused a lot of issue due to it's destructor, also since it's
plan to replace it by std::vector at the end, reduce it's power in the code is
the best.

For the moment logic bricks and armatures contraints/channels stil inherite from
SG_QList but not though PyObjectPlus.