UPBGE: Implement lod python interface.
Previously it was impossible to the user to read lod datas or
change the list of lod, which causes that changing the mesh of an
object which used lod was useless.

To fix it the user should acces to the lod in python. The two existing
classes KX_LodManager (originally KX_LodList) and KX_LodLevel (originally
KX_LodList::Level) now contain a python proxy.

The KX_LodManager is exposed in python in gameobj.lodManager and
KX_LodLevel in lodManager.levels[i].

KX_LodLevel contains the following attributs:
- mesh
- level
- distance
- hysteresis
- useMesh
- useMaterial
- useHysteresis

The behavior of these two classes is still the same. Excepted the new
lod distance factor introduced in KX_LodManager and KX_Camera repectivelly
under the name in python of lodManager.distanceFactor and
camera.lodDistanceFactor. They both values are used to multiply the
distance object to camera to allow the user to set a kind of zoom.

As noticed before the user can now change the lod manage to change the
lod levels list. This change is only done by asignement. To manage
changement the m_previousLodLevel is set to -1 which has as effect in
KX_LodManager::GetLevel to disable hysteresis. If the user set the lod
manager to None then the object's mesh backs to the mesh in the previous
first lod level.