UPBGE: Implement LOD python API
Complete commit message:

- KX_LOD is refactored in 2 files: KX_LodManager (each KX_GameObject has
  its own LodManager that contains a list of all levels of detail) and
  KX_LodLevel that contains the attributes of each level of detail.

- A new python API is created at several levels for more control:

1) KX_LodLevel: you can set the attributes of a single level of detail
(distance, hysteresis) and get infos about this level of detail (the mesh
name used for this level of detail, and if this level of detail use

hysteresis override option.
2) KX_LodManager: you can set the distance of all levels of detail
contained in levels of detail list for a specific KX_GameObject. An
3) KX_Scene: A method setLodScale(PythonObjectList, scale) is implemented.
This way, the user can create his own lists of KX_GameObjects with LOD
(maximum control, maximum force, maxi hamburger) and pass his list of
KX_GameObject to the function and apply a scale to all the list.

If KX_GameObjects are added during runtime, they will share the same
instance of KX_LodManager than the original object but the distance to the
camera of this object won't be the same so there is no problem I think.
When he adds objects (trees for example as I have a lot of imagination),
he can append this objects to its initial list of trees and when he remove
trees he can pop the trees from his list. So he manages his lists as he
wants. The better is to do small lists of course.

So I think that no one will use this API but I was happy to do it
nevertheless.