Implement logic.keyboard.text
This value represent the text typed between two frames. The special value for
return key or delete key are converted in \r and \b respectively.
Which mean that the "text" value should not be printed as it since the \r and \b
effect will be applied.

A bug was found in the same moment, resverse translate table for event was
mixing mouse event type and key event type in the same range. Which causes for
example that delete key was handled after KX_ENDKEY.

Spliting reverse table in 3 tables for keys, mouse and system events solve the
issues as using special convert event function for these 3 event categories.