UPBGE: Render text objects as mesh objects.
Previously all text objects was rendered at the frame render end. It caused that
as text objects use transparency, they were all on top of all other transparent
regular meshes which are not writing to the depth buffer.

To solve this issue the text objects should be rendered like the other regular meshes
and enjoy the ordering made for transparency objects. But this would assume that
text objects are meshes which is not really the case as they don't use any materials.

To first considerate the text objects as meshes they need a material. This material
is particular as it doesn't bind any shader. It exists only to say that meshes
using it are texts. This material must be global to the game engine to make sure
that we use as much as possible a common RAS_DisplayArrayBucket and RAS_MaterialBucket.

The material is then a subclass of RAS_IPolyMaterial named KX_TextMaterial which
contains mostly empty functions, only it's constructor enable material text flags.
An instance of this material is stored statically in KX_TextMaterial.cpp and get its
pointer with GetTextMaterial from KX_TextMaterial.h.

RAS_IPolyMaterial contains a new function IsText which returns true for KX_TextMaterial
depending of a text flag. This function is used when creating the material bucket to
add the material bucket in a specific list accessed with m_buckets[TEXT_BUCKET].
The function is used again when rendering to call IndexPrimitivesText with the mesh
user.

As texts objects doesn't have any RAS_MeshObject and RAS_DisplayArray, then the
RAS_DisplayArrayBucket must check if the m_mesh is non NULL before for example get if
it is modified. Also m_mesh is used to differ a modifier display array bucket than a
text display array bucket in FindDisplayArrayBucket as the both have a NULL display
array.

As KX_FontObject inherit of KX_GameObject, we just have to make virtual functions
managing mesh user and mesh slots (AddMeshUser, UpdateBuckets). The override of
AddMeshUser just find or create the material bucket with the global KX_TextMaterial.
UpdateBuckets updates all data needed to render the texts in the RAS_TextUser.

The RAS_TextUser is a subclass of RAS_MeshUser containing extra data used to render
the texts:
- dpi
- size
- aspect
- resolution
- fontid
- offset
- spacing
- texts[]

In IndexPrimitivesText we loop on all text contained in RAS_TextUser::m_texts, translate
text matrix and call RenderText3D with all other values contained in RAS_TextUser.

Two bugs are fixed with this patch coresponding to issues #119 and #114. #119 is about
unordered transparency and #114 is for text rendered even if they are inactive because
the scene render all object in the list of texts which is wrong as some of them can be
inactive. This is fixed by the fact that only active objects update theirs mesh user and
add theirs mesh slot to be rendered.

This patch allow for the moment to draw text objects as meshes with an empty material,
but later if a way is found to render texts with original material, the rest will be
already done.