Fix compatibility for if event:
This translate the test to:
if KX_INPUT_ACTIVE in event.status:

But if the user want to know that a variable event is not None he have to
do it explicitly :
if event is not None.

To summarize the user now must be explicit in any test with SCA_InputEvent.