UPBGE: Implement game object bounding volume box python API.
The user wasn't able to set or just read the AABB previously, this can be useful
for "predefined AABB" for animated objects or at least to know the volume of the
mesh.

The API now exposes a KX_BoundingBox python class accesible in KX_GameObject.cullingBox.
This class is temporary created when accesing obj.cullingBox, none logic is proceed in
it for internal usage.
The class contains the following attributes:
- min (Vector3, rw),
- max (Vector3, rw),
- center (Vector3, ro),
- radius (float, ro),
- autoUpdate (bool, rw)