UPBGE: Implement common launcher system and common input event system.
Previously two major issues in the (UP)BGE was that the launcher system
was not common which created a lot of duplicated code between the blender
player and the embedded player.
The second issue was about input event, the events gives to the user was
the events handled only at the begining of the frame. For example if the
game run a 5fps and that the user press a key quickly the kye event will
be not handled.

Input Event System:

First as the embedded input system used wmWin* struct and classes which
used GHOST, a common input system is created based only on GHOST.
These classes are put in the directory Ghost/, it contains the input device
GH_InputDevice which convert every GHOST event in SCA input enums, and
the event consumer GH_EventConsumer.

After that the SCA_InputEvent had to be changed to expose to the user every
status, events and values handled from the last frame. To do it this class
now use a python proxy and it is accesible in python from:

bge.logic.keyboard.inputs[enum]

And in SCA_InputEvent python proxy:

event.status
event.queue
event.values
event.type

All these attributs are explained precislly in the python documentation.

For compatibility and naming issue [keyboard/sensor].[[active_events]/events]
are left but deprecated to use [keyboard/sensor].[[activeInputs]/inputs].

Common Launcher System:

As the input system is now shared it's easier to share the launcher system.
The launcher process is refactor in a suite of class inspired of GPG_Application.

They are:
LA_Launcher:
(1)LA_BlenderLauncher
(2)LA_PlayerLauncher

LA_Launcher contains the majority of the code and the both subclasse for the
blender player and the embedded player are used for specific initialization
as call GPU_init for blenderplayer or set the proper canvas dimensions.

All these classes are in the directory Launcher/.

The LA_Launcher contains three major functions:
InitEngine: Start the engine
ExitEngine: Stop the engine
EngineMainLoop: Setup a while loop of frame call

In the same moment the opened file name acces was changed to a global variable:
KX_GetMainPath: return the name of the current opened file
KX_GetOrigPath: return the value of G.main->name at the game start

By this way we avoid modify the G.main->name as before.