UPBGE: Fix reload or start other game file.
Previouslly the file was properly loaded but the scene converted was always
the first one from the original file. This switched to the current scene
in the loaded file.

The only side effect of this fix is that the setting back to the loaded scene
and not the original scene as before.