UPBGE: Implement soft particles/depth transparency.
Soft particles, or depth transparency, is a technique used to blend
transparent meshes (planes in general) with the other meshes behind
depending of the depth between the two.

For the user it's a very simple interface: An option in the UI to enable
depth transparency and a factor parameter to know how much blending
the meshes with the others.

To implement this feature we first needed to access to the depth texture,
it's why this branch waited for off screen render.
After we have to pass the depth texture to the shader. It wasn't possible
because the depth texture will change for each windows resizing and per
game start. To do it we need to use a pointer to this depth texture
(GPUTexture **). The function GPU_dynamic_texture_ptr does this job.

Multisample render off screen need to be blit to sample the depth in
the shader. To allow this mechanisme a new off screen type is added:
RAS_OFFSCREEN_BLIT_DEPTH. Also two paramteter are added to
GPU_offscreen_blit to choose if the color or the depth or the both will
be blit.

At the end the shader need to read the depth and compare it by interpolating
the current vertex position and the vertex position scale, with the
depth value as factor.