UPBGE: Billboard optimization and object rotation.
Previously the user was not able to set a billboard axis and the axis as default was
the Z axis. Using different axis than Z can be useful in some case; e.g the flame of
a gun aligned in X or Y axis. The user is now able to set an object rotation which
will simulate an axis of the billboard. The object Z axis is used as the billboard axis.
This rotation is used for both billboard and halo face orientation.

The GetTransform function in case of billboard and halo was optimized by different
ways. Before the patch, for 10000 wired billboard plane the rasterizer time was 3.94ms.
By avoid copying dir to left and other copy operation we are now at 3.62ms. By using
constants values the time down to 3.56ms. And finally if we compute the axis scale
from the matrix and avoid a cast of client object to KX_GameObject and a call to
NodeGetWorldScale() the time go to 3.40ms.