UPBGE: Implement variable vertex memory format.
Previously the vertex class (RAS_TexVert) was using 8 channel for UV (64bytes).
But the user use 8 channel very rarely. In the most of case only the first was
used.

The issue is the memory used for a vertex which is too high and will be more
for futur multiple color channels.

To reduce the memory size we have to make a vertex class with a UV channel number
as template argument to size the uv array to the uv channels used.

RAS_TexVert<unsigned int UVSize>

This class is a subclass of the base class RAS_ITexVert for multiple reasons:
- Keep using vertex class in deformers and avoid using indirect virtual
  RAS_DisplayArray function.
- Allow base functions for getting/setting position/normal/tangent to be inlined.

In this inheritance as all base functions are inlined only the following functions
are virtual:

virtual const unsigned short getUVSize() const
virtual const float *getUV(const int unit) const
virtual void SetUV(const int index, const MT_Vector2& uv)
virtual void SetUV(const int index, const float uv[2])

The only disadvantage of an inheritance is that the vptr take 8 bits more in the
vertex class.
To counter this side effect, some vertex data are split in a new class RAS_TexVertInfo.
The data contained are:

int m_origindex
short m_softBodyIndex
short m_flag

The reason to doing this is that these data are not needed for the render and are
very rarely accessed (only for deformers).

About RAS_DisplayArray a similar change to RAS_TexVert was applied. RAS_DisplayArray
take as template parameter the vertex class. It contains the std::vector of vertexes
and it is based upon RAS_IDisplayArray.

This class is now obligated to give access to vertexes through a virtual function,
But as this function is virtual calling it spend a lot of time. To optimize the access
to vertexes a cache system is setup. RAS_IDisplayArray now contains a list named
m_vertexPtrs, this list contains the pointer of all vertexes at the end of conversion.
Then calling:

inline RAS_ITexVert *GetVertex(const unsigned int index) const

Return a pointer directly from the cache list. The disadvantages is the memory consumed
(8 byte per vertex) and the UpdateCache function needed to be called at the end of
conversion or after duplication.

The cache cannot be used during the conversion, then a true virtual function is used:

virtual RAS_ITexVert *GetVertexNoCache(const unsigned int index) const

As the vertex type is dynamic we have to pass the template paramter to the functions/classes
which instantiate the RAS_DisplayArray. A new class RAS_TexVertFormat, which is currently
only containing the uv size, is passed to the functions:

RAS_MeshObject::AddMaterial
RAS_MaterialBucket::AddMesh
RAS_MeshSlot::init

The uv size value is choose in BL_ConvertMesh with the expression:
max_ii(1, validLayers)

About the rasterizer storages, the modification are pretty minors. Only VBO need to
get offset in an other way, because computing the offset with a virtual function
(getUV) with a NULL instance raise a segfault. So RAS_DisplayArray gives offset with
functions:

virtual unsigned int GetVertexMemorySize() const
virtual void *GetVertexXYZOffset() const
virtual void *GetVertexNormalOffset() const
virtual void *GetVertexTangentOffset() const
virtual void *GetVertexUVOffset() const
virtual void *GetVertexColorOffset() const

At the end, variable memory vertexes allow optimization for animations and soft body as
manipuling a small piece of meory is faster.