UPBGE: Implement realtime cube map rendering.
The idea of realtime cube map rendering is to render each faces
of the cube with the proper camera orientation. Then this cube
map textures is used by materials for reflection effects.
These reflections in realtime can be used in a car game or for
some soft bodies.

The usage of cube map is contrained of one realtime cube map per
texture, not object neither material texture.

The interface to the user is in in two sides: an UI and a python
API.

About the UI, a new environment map source type was added: "REALTIME".
This mode offers the same option as the mode "STATIC" or "ANIMATED"
exepted that the texture resolution and depth is removed. Also this
source mode request to use an exisiting image.

Two UI options was added: "Auto Update" and "Filtering".
The first option allow to update, or not, the realtime cube map
for every frame. The second option is the texutre filtering methode
used, one of:
- None: No filtering
- Linear Filtering: Texture filtering without mipmapping
- Mipmap Filtering: Texture filtering with mipmapping generation

The image adding from the texture now raise an error when the image
size is invalid for a cube map. At least a image is created with
the maximum valid size corresponding to the user inputs

From the python API side, a python proxy class is added: KX_CubeMap.
This class is accessed from BL_Texture.cubeMap. An instance of
KX_CubeMap can be shared bewteen BL_Texture's as this last class
represent a material texture not a single texture.
The KX_CubeMap class contains the following attributes and functions:
- clipStart
- clipEnd
- autoUpdate
- ignoreLayers
- enabled
- viewpointObject
- update()

All the attributes rely on the UI options exepted for the only
functions "update()". THis function ask to render the cube map for
the nest rendering stage.

In an other world, the sources, some changement was made around the
viewport gpu part and the game engine rasterizer part.

In the gpu part, the frame buffer APi chenged to accept using a cube
map attachement, e.g GL_TEXTURE_CUBE_MAP_POSITIVE_Z_ARB.

To do this a new function is added:
int GPU_framebuffer_texture_attach_target(GPUFrameBuffer *fb, struct
GPUTexture *tex, int target, int slot, char err_out[256]);

where target is the OpenGL target enumeration.

In the same time a function do detach a render buffer objects like
for textures is implemented:
void GPU_framebuffer_renderbuffer_detach(GPURenderBuffer *rb);

About the game engine, a RAS_CubeMap and KX_CubeMap class are added.
RAS_CubeMap is managing the frame buffer objects and render buffer
objects creation attaching the textures and the render buffer objects.
There's 6 FBO's and RBO's for the 6 faces.

KX_CubeMap is a subclass of RAS_CubeMap, it contains the python proxy
and all value of the environment map not requested by the FBO's, e.g:
ignoreLayers, autoUpdate.

Finally a manager class KX_CubeMapManager owned by the scene proceed
the rendering. From outside the function Render is called and
internally RenderCubeMap is called for every realtime cube maps.

This class also manage the invalidation of realtime cube maps when
a viewpoint object is freed through the function
InvalidateCubeMapViewpoint.

The rendering of each faces first set the projection matrix of the
cube map and the modelview matrix of the rendered face, the camera
is updated with the same datas.
Then culling is proceed same as animations. After this the render is
made.